update styled print class
